fix: Reduce number of optional fields to catch API changes
The names of the fields for distance and area were changed without
being documented or the API version changed. To try and catch changes
like this, the code now checks the protocol and sets the appropriate
fields, rather than treating them as optional.

Also, values for distance and area are returned from the API as
floats, so to handle any precision issues, the decimals are now
rounded to three places.